Добавил журналирование переводов между страхователями, дополнил журналирование платежей на целевой счет, добавил журналирование на пополнение (надо доработать)

+++ src/main/java/beans/Account.java
@@ -26,7 +26,7 @@ public class Account {
     /**
      * Сумма на счете
      */
-    private int sum;
+    private double sum;
 
     /**
      * Конструктор счета по его номеру
