+++ src/main/java/processors/AccountProcessor.java
@@ -11,7 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class AccountProcessor {
 
-    private static void validateReplenishSum(int replenishSum) throws ReplenishException{
+    private static void validateReplenishSum(double replenishSum) throws ReplenishException{
         log.info("Валидация суммы пополнения счета");
         if(replenishSum <= 0){
             log.error("Ошибка при пополнении - сумма пополнения меньше или равна нулю");
@@ -25,14 +25,14 @@ public class AccountProcessor {
      * @param replenishSum сумма пополнения
      * @return флаг пополнения счета
      */
-    public static synchronized void replenishAccount(Account account, int replenishSum) throws ReplenishException{
+    public static synchronized void replenishAccount(Account account, double replenishSum) throws ReplenishException{
         log.info(String.join(" ", "Пополнеие счета", String.valueOf(account.getAccountNumber()), "на сумму ", String.valueOf(replenishSum)));
         validateReplenishSum(replenishSum);
         account.setSum(account.getSum() + replenishSum);
         log.info(String.join(" ", "Пополнение счета", String.valueOf(account.getAccountNumber()), "прошло успешно. Текущая сумма на счете", String.valueOf(account.getSum())));
     }
 
-    private static boolean validateWithdrawalSum(int accountSum, int withdrawalSum) throws WithdrawalException{
+    private static boolean validateWithdrawalSum(double accountSum, double withdrawalSum) throws WithdrawalException{
         log.info("Валидация суммы списания со счета");
         if(withdrawalSum <= 0){
             log.error("Ошибка при списании - сумма списания меньше или равна нулю");
@@ -50,7 +50,7 @@ public class AccountProcessor {
      * @param withdrawaslSum
      * @throws WithdrawalException
      */
-    public static void withdrawalAccount(Account account, int withdrawaslSum) throws WithdrawalException{
+    public static void withdrawalAccount(Account account, double withdrawaslSum) throws WithdrawalException{
         log.info(String.join(" ", "Списание со счета", String.valueOf(account.getAccountNumber()), "на сумму", String.valueOf(withdrawaslSum)));
         validateWithdrawalSum(account.getSum(), withdrawaslSum);
         account.setSum(account.getSum() - withdrawaslSum);
@@ -65,7 +65,7 @@ public class AccountProcessor {
      * @throws WithdrawalException
      * @throws ReplenishException
      */
-    public static void transferMoney(Account fromAccount, Account toAccount, int transferSum) throws WithdrawalException, ReplenishException{
+    public static void transferMoney(Account fromAccount, Account toAccount, double transferSum) throws WithdrawalException, ReplenishException{
         log.info(String.join(" ", "Перевод средств со счета",
                 String.valueOf(fromAccount.getAccountNumber()), "на счет",
                 String.valueOf(toAccount.getAccountNumber()), "на сумму", String.valueOf(transferSum)));
